style: move formatDate to common-utils
